+++ src/commands/images/buildImage.ts
@@ -9,8 +9,8 @@ import { IActionContext } from "vscode-azureextensionui";
 import { ext } from "../../extensionVariables";
 import { localize } from '../../localize';
 import { getOfficialBuildTaskForDockerfile } from "../../tasks/TaskHelper";
-import { delay } from "../../utils/delay";
 import { getValidImageName } from "../../utils/getValidImageName";
+import { delay } from "../../utils/promiseUtils";
 import { quickPickDockerFileItem } from "../../utils/quickPickFile";
 import { quickPickWorkspaceFolder } from "../../utils/quickPickWorkspaceFolder";
 import { selectBuildCommand } from "../selectCommandTemplate";
